requirements.txt
streamlit==1.45.0
pandas==2.2.3
pillow==11.2.1
requests==2.31.0
e2b-code-interpreter==1.5.0
python-dotenv==1.0.0
